requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.54
Flask-Migrate==4.0.4
python-dotenv==1.0.1
PyMySQL==1.1.1
